script/pdf to ppt/requirements.txt
pdf2image>=3.1.0
python-pptx>=0.6.21
Pillow>=9.0.0
